Add create user test to test.ts
