+++ package.json
@@ -9,7 +9,7 @@
     "dev": "tsx watch ./src/index.ts",
     "lint": "eslint ./src --ext .ts",
     "build": "tsc",
-    "test": "mocha -r tsx test/test.ts"
+    "test": "DOTENV_CONFIG_PATH=./.test.env mocha -r tsx -r dotenv/config src/test/test.ts"
   },
   "keywords": [],
   "author": "",
@@ -26,6 +26,7 @@
     "@types/node": "22.15.14",
     "axios": "1.9.0",
     "chai": "5.2.0",
+    "dotenv": "16.5.0",
     "eslint": "9.26.0",
     "eslint-config-prettier": "10.1.2",
     "eslint-plugin-prettier": "5.4.0",
